Przed zmianą filtrów

+++ src/app/auth/login/login.component.ts
@@ -1,11 +1,8 @@
 import { Component, OnDestroy, OnInit } from '@angular/core';
-import { NgForm } from '@angular/forms';
 import { AngularFireAuth } from '@angular/fire/auth';
-import { MatSnackBar } from '@angular/material/snack-bar';
 import { AuthService } from '../auth.service';
-import { Router } from '@angular/router';
 import { firebase, firebaseui, FirebaseUIModule } from 'firebaseui-angular';
-import { BehaviorSubject } from 'rxjs';
+import { Subscription } from 'rxjs';
 
 @Component({
   selector: 'app-login',
@@ -15,17 +12,15 @@ import { BehaviorSubject } from 'rxjs';
 export class LoginComponent implements OnInit, OnDestroy {
   hide = true;
   ui: firebaseui.auth.AuthUI;
+  isLoading: boolean;
+  sub: Subscription;
 
-  constructor(
-    private afAuth: AngularFireAuth,
-    private _snackBar: MatSnackBar,
-    private authService: AuthService,
-    private router: Router
-  ) {}
+  constructor(private afAuth: AngularFireAuth) {}
 
   ngOnInit(): void {
+    this.isLoading = true;
     const firebaseUiAuthConfig: firebaseui.auth.Config = {
-      // signInFlow: 'popup',
+      signInFlow: 'popup',
       signInOptions: [
         firebase.auth.GoogleAuthProvider.PROVIDER_ID,
         {
@@ -34,53 +29,27 @@ export class LoginComponent implements OnInit, OnDestroy {
         },
         firebaseui.auth.AnonymousAuthProvider.PROVIDER_ID,
       ],
-      callbacks: {
-        signInSuccessWithAuthResult: this.onLoginSuccesful.bind(this),
-      },
     };
 
-    this.afAuth.app.then((app) => {
-      this.ui = new firebaseui.auth.AuthUI(app.auth());
-      this.ui.start('#firebaseui-auth-container', firebaseUiAuthConfig);
+    this.sub = this.afAuth.authState.subscribe((user) => {
+      if (user) {
+        this.isLoading = true;
+      } else {
+        this.isLoading = false;
+        this.afAuth.app.then((app) => {
+          this.ui = new firebaseui.auth.AuthUI(app.auth());
+          this.ui.start('#firebaseui-auth-container', firebaseUiAuthConfig);
+        });
+      }
     });
-
-  }
-
-  onLoginSuccesful(result) {
-    console.log(result)
   }
 
   ngOnDestroy() {
-    this.ui.delete();
+    if (this.ui) {
+      this.ui.delete();
+    }
+    if (this.sub) {
+      this.sub.unsubscribe();
+    }
   }
-
-  // onSubmit(form: NgForm) {
-  //   this.firebase
-  //     .signInWithEmailAndPassword(form.value.email, form.value.password)
-  //     .then(
-  //       (success) => {},
-  //       (error) => {
-  //         this._snackBar.open(error.message, 'Ok', {
-  //           duration: 10000,
-  //           verticalPosition: 'top',
-  //         });
-  //       }
-  //     );
-  //   form.reset();
-  // }
-
-  // logWithNoAccount() {
-  //   this.firebase.signInAnonymously().then(
-  //     (success) => {
-  //       this.authService.userId.next(success.user.uid);
-  //       this.router.navigate(['/movies']);
-  //     },
-  //     (error) => {
-  //       this._snackBar.open(error.message, 'Ok', {
-  //         duration: 10000,
-  //         verticalPosition: 'top',
-  //       });
-  //     }
-  //   );
-  // }
 }
